finished welcome page with login and registration
