+++ api/database.json
@@ -1,0 +1,48 @@
+{
+    "users":  [
+        {"id": 1,
+        "name": "Herman Munson",
+        "email": "herm@adamshouse.com"
+        },
+    {
+        "id": 2,
+        "name": "Lily Munson",
+        "email": "lily@adamshouse"
+    }
+    ],
+    "friends": [
+     {
+        "id": 1,
+        "users_id": 1,
+        "users_id_fk": 2
+     }
+    ],
+    "events": [
+        {
+        "id": 1,
+        "title": "Party at the Moon Tower",
+        "date": "now",
+        "time": "now",
+        "location": "Moon Tower"
+        }
+    ],
+    "news": [
+        {
+        "id": 1,
+        "url": "https://www.imdb.com/title/tt0387808/",
+        "title": "End of the World",
+        "synopsis": "Pestilence, War, Famine, Death"
+        }
+
+
+    ],
+    "tasks": [
+        {
+            "id": 1,
+            "task": "clean litter box",
+            "date": "now",
+            "dateCompleted": "now, now"
+        }
+
+    ]
+}
